fixing warnings
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@5916 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/traffic_lights/MSTrafficLightLogic.h
@@ -93,7 +93,7 @@ public:
     virtual size_t getPosition(SUMOTime simStep) = 0;
 
     /// returns the step (the phasenumber) of a given position of the cycle
-    virtual size_t getStepFromPos(size_t position) = 0;
+    virtual unsigned int getStepFromPos(unsigned int position) = 0;
 
     /// Returns the phase of a given step
     virtual const MSPhaseDefinition &getPhaseFromStep(size_t givenstep) const = 0;
